requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
datasets==2.4.0
dill==0.3.5.1
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.7.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.7.3
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.12.1
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1